src/requirements.txt
Flask
Flask-Migrate
connexion[swagger-ui]
sqlalchemy
urllib3
Werkzeug
alembic
marshmallow
marshmallow-sqlalchemy
psycopg2
pyjwt
netifaces
python-consul